feat(server): initialize nest server code
nest g application server ran in root to create server project. Also, dependencies of server moved
to main package.json so the dist folder can be ran with the required depndencies. the client side
don't need that because the client side files compile all dependencies in the final js file